requirements.txt
ag-ui-protocol==0.1.18
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
argcomplete==3.6.3
asttokens==3.0.1
attrs==26.1.0
Authlib==1.7.0
bcrypt==5.0.0
beartype==0.22.9
blinker==1.9.0
boto3==1.43.0
botocore==1.43.0
brotli==1.2.0
build==1.4.4
cachetools==7.0.6
caio==0.9.25
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
cohere==5.21.1
contourpy==1.3.3
cryptography==47.0.0
cssselect2==0.9.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.4
cuda-toolkit==13.0.2
cycler==0.12.1
cyclopts==4.11.0
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.18.0
docutils==0.22.4
durationpy==0.10
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastavro==1.12.2
fastmcp==3.2.4
filelock==3.29.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
genai-prices==0.0.57
gitdb==4.0.12
GitPython==3.1.49
google-auth==2.49.2
google-genai==1.74.0
googleapis-common-protos==1.74.0
graphviz==0.21
griffelib==2.0.2
groq==1.2.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.2
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
joserfc==1.6.4
jsonpath-python==1.1.5
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.6
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.5.0
kubernetes==35.0.0
logfire==4.32.1
logfire-api==4.32.1
logging==0.4.9.6
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
matplotlib-inline==0.2.1
mcp==1.27.0
mdurl==0.1.2
mistralai==2.4.4
mmh3==5.2.1
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
narwhals==2.20.0
networkx==3.6.1
nexus-rpc==1.4.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.11.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.26.2
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnxruntime==1.25.1
openai==2.33.0
openapi-pydantic==0.5.1
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.8
overrides==7.7.0
packaging==25.0
pandas==3.0.2
parso==0.8.6
pathable==0.5.0
patsy==1.0.2
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.4.4
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.3
pydantic-ai==1.88.0
pydantic-ai-slim==1.88.0
pydantic-evals==1.88.0
pydantic-graph==1.88.0
pydantic-handlebars==0.1.0
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydeck==0.9.2
pydyf==0.12.1
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pyperclip==1.11.0
pyphen==0.17.2
PyPika==0.51.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.17.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.5.0
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sse-starlette==3.4.1
stack-data==0.6.3
starlette==1.0.0
statsmodels==0.14.6
streamlit==1.57.0
sympy==1.14.0
tabulate==0.10.0
temporalio==1.26.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tk==0.1.0
tokenizers==0.22.2
toml==0.10.2
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.3
traitlets==5.14.3
transformers==5.7.0
triton==3.3.0
typer==0.25.0
types-protobuf==6.32.1.20260221
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.3.1
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xai-sdk==1.12.0
yarl==1.23.0
zipp==3.23.1
zopfli==0.4.1
